+++ cider-eval.el
@@ -806,6 +806,16 @@ buffer."
                             nil
                             (cider--nrepl-pr-request-map))))
 
+(defun cider-eval-list-at-point (&optional output-to-current-buffer)
+  "Evaluate the list (eg. a function call, surrounded by parens) around point.
+If invoked with OUTPUT-TO-CURRENT-BUFFER, output the result to current buffer.
+Special cases such as deref-ing a function call's results is also executed,
+like in @(fn-that-returns-an-atom x)"
+  (interactive "P")
+  (save-excursion
+    (goto-char (cadr (cider-list-at-point 'bounds)))
+    (cider-eval-last-sexp output-to-current-buffer)))
+
 (defun cider-eval-sexp-at-point (&optional output-to-current-buffer)
   "Evaluate the expression around point.
 If invoked with OUTPUT-TO-CURRENT-BUFFER, output the result to current buffer."